chore: cypress e2e config
