+++ cypress.config.ts
@@ -9,9 +9,10 @@ export default defineConfig({
   },
 
   e2e: {
-    baseUrl: 'http://localhost:8081',
+    baseUrl: 'http://localhost:6006',
     setupNodeEvents(on, config) {
       // implement node event listeners here
     },
+    specPattern: 'cypress/integrations/**/*.cy.{ts,tsx}',
   },
 });
