+++ cypress/support/commands.ts
@@ -37,3 +37,8 @@ import '@testing-library/cypress/add-commands'
 //     }
 //   }
 // }
+
+Cypress.Commands.add('visitStory', (storyName, options) => {
+  // eslint-disable-next-line cypress/no-unnecessary-waiting
+  return cy.visit(`iframe.html?id=${storyName}`, options).wait(0);
+});
